% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{build_link_table}
\alias{build_link_table}
\title{Build the link table from the initial tabulation table (data.frame or data.table) where each row
correspond to the number of statistical units in a cross defined by a modality of the z1 nomenclature and a modality of the z2 nomenclature. 
One line of the unit table represents a connection, triplets composed of 2 elements of z1 and one element of z2 connecting. 
The identifier of connected component in the graph of the elements of z2 is added as well as the number of observations contained 
in the intersection between the zone of z2 connecting and the 2 elements of z1 connected by z2}
\usage{
build_link_table(input_df)
}
\arguments{
\item{input_df}{The tabulation table (data.frame or data.table). Each row
corresponds to the number of statistical unitsin a cross defined by a modality of the z1 nomenclature and a modality of the z2 nomenclature}
}
\value{
the link_table containing the triplet and metadatas (data.table)
}
\description{
Build the link table from the initial tabulation table (data.frame or data.table) where each row
correspond to the number of statistical units in a cross defined by a modality of the z1 nomenclature and a modality of the z2 nomenclature. 
One line of the unit table represents a connection, triplets composed of 2 elements of z1 and one element of z2 connecting. 
The identifier of connected component in the graph of the elements of z2 is added as well as the number of observations contained 
in the intersection between the zone of z2 connecting and the 2 elements of z1 connected by z2
}
\examples{
link_table<- build_link_table(input_df)
}
