% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{build_m_crois}
\alias{build_m_crois}
\title{Build the cross matrix}
\usage{
build_m_crois(link_table)
}
\arguments{
\item{link_table}{(data.table) containing the triplets z1-z2-z1 corresponding to a connection between 2 elements of z1 throug one element of z2}
}
\value{
returns the crossover matrix (SparseMatrix), whose rows represent the elements of z1 and the columns the elements of z2 (possibly aggregated if they contribute to the same connection) and whose values are equal to the number of statistical units in the considered z1 x z2 crossover
}
\description{
Build the cross matrix
}
\examples{
m_crois<- build_mcrois(link_table)
}
