% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{clean_init_df}
\alias{clean_init_df}
\title{Cleans the initial dataset by removing rows that are not in the correct format and removes duplicated lines}
\usage{
clean_init_df(input_df)
}
\arguments{
\item{input_df}{The tabulation table (data.frame or data.table). Each row
corresponds to the number of statistical unitsin a cross defined by a modality of the z1 nomenclature and a modality of the z2 nomenclature}
}
\value{
the cleaned tabulation table
}
\description{
Cleans the initial dataset by removing rows that are not in the correct format and removes duplicated lines
}
\examples{
cleaned_df<- clean_init_df(input_df)
}
