% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{clean_init_dt}
\alias{clean_init_dt}
\title{Cleans the initial dataset by removing rows that are not in the correct format
 and aggregates duplicated lines}
\usage{
clean_init_dt(input_dt)
}
\arguments{
\item{input_dt}{The tabulation table (data.table). Each row
corresponds to the number of statistical units in a cross defined by a
modality of the z1 nomenclature and a modality of the z2 nomenclature}
}
\value{
the cleaned tabulation table
}
\description{
Cleans the initial dataset by removing rows that are not in the correct format
 and aggregates duplicated lines
}
\examples{
input_dt<- toy_example_6
cleaned_dt <- diffman:::clean_init_dt(input_dt)
}
