% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{create_fictive_ind_table}
\alias{create_fictive_ind_table}
\title{create the equivalent individual data.frame from a tabulate z1 x z2 
data.frame with integer (round in any case)}
\usage{
create_fictive_ind_table(tab_table)
}
\arguments{
\item{tab_table}{The tabulation table (data.frame or data.table). Each row
corresponds to the number of statistical units in a cross defined by a 
modality of the z1 nomenclature and a modality of the z2 nomenclature}
}
\value{
the fictive individual table from the tabulation table
}
\description{
create the equivalent individual data.frame from a tabulate z1 x z2 
data.frame with integer (round in any case)
}
\examples{
tab_table <- toy_example_1
create_fictive_ind_table(tab_table)

}
