% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging.r
\name{fusion}
\alias{fusion}
\title{Fonction qui fusionne les zones numero i et numero j}
\usage{
fusion(i, j, m_crois, col_to_suppress, verbose = FALSE)
}
\arguments{
\item{i, j}{Entiers indiquant les zonages de z1.}

\item{m_crois}{Matrice de croisement.}

\item{col_to_suppress}{Vecteur d'entiers indiquant les colonnes a supprimer.}

\item{verbose}{if TRUE a message is displayed if every z1 elements habs been merged}
}
\value{
En sortie, on a la matrice de croisement avec une ligne en moins,
et potentiellement des colonnes en moins egalement.
}
\description{
Fusion des zones i et j du zonage z1.
}
\details{
Les colonnes a supprimer \code{col_to_suppress} sont determinees
lors du test de fusion. Si le test conclue qu'il faut fusionner
deux zones, alors on retourne en plus les zones de z2 qui sont
a l'interieur de cette zone fusionnee.
}
