% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{isolate_intersection}
\alias{isolate_intersection}
\title{Isolate intersection z2xz1 from z2 lines where z2 is fully included in z1}
\usage{
isolate_intersection(input_dt)
}
\arguments{
\item{input_dt}{(data.table) containing the z2 x z2 crossing with counts, tabulation table}
}
\value{
a list with the intersection on the first element and the llines corresponding to the fully included square in the second element
}
\description{
Isolate intersection z2xz1 from z2 lines where z2 is fully included in z1
}
