% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{long_table}
\alias{long_table}
\title{Transform the link table containing triplets z1-z2-z1 with metadata about connection into a long table where each line correspond to  a part of the connections z1-z2 and the number of statistical unitis inside the intersection of the corresponding z1 and z2}
\usage{
long_table(link_table)
}
\arguments{
\item{link_table}{(data.table) containing the triplets z1-z2-z1 corresponding to a connection between 2 elements of z1 throug one element of z2}
}
\value{
the long table with the metadata, identifiying the actual connected components in the z1 graph
}
\description{
Transform the link table containing triplets z1-z2-z1 with metadata about connection into a long table where each line correspond to  a part of the connections z1-z2 and the number of statistical unitis inside the intersection of the corresponding z1 and z2
}
\examples{
ltable<- long_table(link_table)
}
