% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{one_component_risk_extraction}
\alias{one_component_risk_extraction}
\title{return the diff info from a input data table corresponding to one connected 
component of z1 perform the search of diff issue and
return the differantiation information from the corresponding component}
\usage{
one_component_risk_extraction(input_dt, threshold, max_agregate_size)
}
\arguments{
\item{input_dt}{The tabulation table (data.frame or data.table). Each row
corresponds to the number of statistical units in a cross defined by a 
modality of the z1 nomenclature and a modality of the z2 nomenclature
z1 has to belong to the same related component
(regarding the graph construction)}

\item{threshold}{Strictly positive integer indicating the confidentiality
threshold. Observations are considered at risk if one can deduce information
on an aggregate of n observations where n < threshold.}

\item{max_agregate_size}{Integer indicating the maximal size of agregates
which are tested exhaustively. If that number is too large (greater than 30),
the computations may not end because of the combinations number
that can become very large. Also the RAM can be overloaded.}
}
\value{
a data.table containing the differentiation info for this component
 (same format than the return diff info function)
}
\description{
return the diff info from a input data table corresponding to one connected 
component of z1 perform the search of diff issue and
return the differantiation information from the corresponding component
}
\examples{
input_dt <- toy_example_6
compo <- return_connected_components(input_dt)
one_component_risk_extraction(input_dt[z1 \%in\% compo[id_comp == 1]$z1], 11, 15)

}
