% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data , cleaning input_df with z1xzé crossin and extract lines containing z2 whichh are not fully contained by one element of z1}
\usage{
prepare_data(input_dt)
}
\arguments{
\item{input_dt}{(data.table) containing the z2 x z2 crossing with counts, tabulation table}
}
\value{
data.table containing only the z2 intersecting 2 z1
}
\description{
Prepare data , cleaning input_df with z1xzé crossin and extract lines containing z2 whichh are not fully contained by one element of z1
}
