% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{read_diff_info}
\alias{read_diff_info}
\title{Read diff_info info from files into a given directory and rbind them}
\usage{
read_diff_info(save_dir)
}
\arguments{
\item{save_dir}{the directory where the diff_info files (from all_component_risk_extraction) are included}
}
\value{
a data.table containning all the diff info for all related components
}
\description{
Read diff_info info from files into a given directory and rbind them
}
