% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_1}
\alias{toy_example_1}
\title{Minimal example with only 2 z1 elements and 1 z2 elements with differencing issue for a treshold equals to 11}
\description{
Minimal example with only 2 z1 elements and 1 z2 elements with differencing issue for a treshold equals to 11
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
