% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_3}
\alias{toy_example_3}
\title{Minimal example with 5 z1 elements, M1-M2 are connected, als o M4-M5 and M3 is isolated}
\description{
Minimal example with 5 z1 elements, M1-M2 are connected, als o M4-M5 and M3 is isolated
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
