% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_4}
\alias{toy_example_4}
\title{Minimal example with 5 z1 elements and 1 z2 element intersecting 3 z1 elements}
\description{
Minimal example with 5 z1 elements and 1 z2 element intersecting 3 z1 elements
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
