% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_5}
\alias{toy_example_5}
\title{Minimal from one connected components extract from french census data}
\description{
Minimal from one connected components extract from french census data
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
