% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_6}
\alias{toy_example_6}
\title{Minimal example with float values, 2 connected compônents M1-M2-M3 and M4-M5}
\description{
Minimal example with float values, 2 connected compônents M1-M2-M3 and M4-M5
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
