% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_7}
\alias{toy_example_7}
\title{Minimal example with float values  one big component M1-M2-M3-MM4-M5-M6}
\description{
Minimal example with float values  one big component M1-M2-M3-MM4-M5-M6
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
