% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.r
\docType{data}
\name{toy_example_8}
\alias{toy_example_8}
\title{Minimal example used in order to illustrated the protection strategy}
\description{
Minimal example used in order to illustrated the protection strategy
}
\author{
Clement Guillo \email{clement.guillo@insee.fr}
}
\keyword{data}
