% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{return_connected_components}
\alias{return_connected_components}
\title{Return elements of z1 graph nodes with the associated connected components}
\usage{
return_connected_components(link_table)
}
\arguments{
\item{link_table}{(data.table) containing the triplets z1-z2-z1 corresponding to a connection between 2 elements of z1 throug one element of z2}
}
\value{
the node_table, each a line corresponds to an element of z1 with the identifier of the connected component to which it belongs in the graph of elements of z1
}
\description{
Return elements of z1 graph nodes with the associated connected components
}
\examples{
link_table<- return_connected_components(link_table)
}
