% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_clement.R
\name{return_diff_info}
\alias{return_diff_info}
\title{Return information on at-risk-of-differenciation z1 zones
From a given list of elements of z1, outputs all the information allowing to
evaluate if a differentiation problem exists on this zone 
(no information if no differneciation issue)}
\usage{
return_diff_info(list_z1, input_dt, threshold = 11)
}
\arguments{
\item{list_z1}{vector containing the elements of z1 constituting the area 
to be evaluated}

\item{input_dt}{(data.table) containing the triplets z1-z2-z1 corresponding 
to a connection between 2 elements of z1 throug one element of z2 with metadata}

\item{threshold}{(data.table) the frequency rule threshold
below which a zone is considered at risk}
}
\value{
A data.table with the diff information
}
\description{
Return information on at-risk-of-differenciation z1 zones
From a given list of elements of z1, outputs all the information allowing to
evaluate if a differentiation problem exists on this zone 
(no information if no differneciation issue)
}
\examples{
list_z1 <- c("M1","M2",c("M1","M2"))
input_dt <- toy_example_1
return_diff_info(list_z1,input_dt)

}
